requirements.txt
keras
tensorflow-gpu
matplotlib
Pillow
opencv-python
sklearn
redis
flask